bot/requirements.txt
aiogram==3.1.1
python-dotenv==1.0.0
aiohttp==3.10.11
requests==2.31.0